refactor: server projects and their backups leave App
Answering a fair question - is App separable, or is it just a knot? Measured rather than
guessed: 3860 lines of logic, 99 pieces of state, 86 refs, 166 top-level functions, and
nine distinct concerns. Not spaghetti - the pure logic is already out and the comments
are alive - but one function doing nine jobs, which is the thing that makes it tiring to
maintain.

So the question is not "can a concern leave" but "which cut is narrowest". I measured
the seam of four candidates - how many App names each reads, and how many of its own the
rest of App reads back - and server storage won: it reads eight things and nothing
outside it reads its state except the JSX that renders its dialogs.

Two hundred lines that never touch cuts, layers, strokes, the canvas or the timeline.
Not the biggest concern in App; the one with the narrowest cut.

buildData and restore stay in App and arrive as two functions. Building a document reads
most of App's state and restoring one writes most of it, so threading either in would
have made the seam wider than the thing it separates.

apiFetch and fetchAsset went to core/api.js first, because App's restore needs them too
and neither owns them: no state, only fetch and a status code. putAsset joined them,
collapsing two upload loops that differed only in their error message.

App.jsx 4324 -> 4158 lines, and the count that matters more: 99 state -> 94, 86 refs ->
80, 166 top-level functions -> 153.

Verification, because this is a change that must alter nothing:

- diffed the moved code against the original. Only five differences, every one intended:
  the two fetch wrappers leaving, the two upload loops becoming putAsset, the two-line
  ref reset becoming forgetProject, a type annotation, and 5 * 60 * 1000 becoming a
  named constant. Everything else is byte-identical.
- eight tests for core/api.js against a real http server, including the 404 that must
  throw rather than come back as a Blob, the upload message surviving so it can be
  numbered, and the extension being escaped into the query.
- the built bundle boots with no console errors and the API probe still answers.
